backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.41
aiosqlite==0.21.0
asyncpg==0.30.0
pydantic==2.11.3
pydantic-settings==2.9.1
python-multipart==0.0.20
redis==5.2.1
openai==1.75.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
numpy==2.2.4
faiss-cpu==1.10.0
pypdf==5.4.0
pillow==11.2.1
pytesseract==0.3.13
matplotlib==3.10.1
boto3==1.38.3
celery==5.5.1
pytest==8.3.5
